+++ typespec-extension/src/models.ts
@@ -1,16 +1,4 @@
-import { Parameter, Schema } from "@autorest/codemodel";
-
-export class LongRunningMetadata {
-  longRunning: boolean = false;
-  pollResultType?: Schema;
-  finalResultType?: Schema;
-
-  constructor(longRunning: boolean, pollResultType?: Schema, finalResultType?: Schema) {
-    this.longRunning = longRunning;
-    this.pollResultType = pollResultType;
-    this.finalResultType = finalResultType;
-  }
-}
+import { Parameter } from "@autorest/codemodel";
 
 export class ClientContext {
   baseUri: string;