+++ typespec-tests/src/main/java/com/_specs_/azure/core/lro/rpc/RpcClient.java
@@ -56,7 +56,18 @@ public final class RpcClient {
      *     status: String(InProgress/Succeeded/Failed/Canceled) (Required)
      *     errors (Optional): [
      *          (Optional){
-     *             error: ResponseError (Required)
+     *             error (Required): {
+     *                 code: String (Required)
+     *                 message: String (Required)
+     *                 target: String (Optional)
+     *                 details (Required): [
+     *                     (recursive schema, see above)
+     *                 ]
+     *                 innererror (Optional): {
+     *                     code: String (Required)
+     *                     innererror (Optional): (recursive schema, see innererror above)
+     *                 }
+     *             }
      *         }
      *     ]
      *     results (Required): [
@@ -91,7 +102,18 @@ public final class RpcClient {
      *     status: String(InProgress/Succeeded/Failed/Canceled) (Required)
      *     errors (Optional): [
      *          (Optional){
-     *             error: ResponseError (Required)
+     *             error (Required): {
+     *                 code: String (Required)
+     *                 message: String (Required)
+     *                 target: String (Optional)
+     *                 details (Required): [
+     *                     (recursive schema, see above)
+     *                 ]
+     *                 innererror (Optional): {
+     *                     code: String (Required)
+     *                     innererror (Optional): (recursive schema, see innererror above)
+     *                 }
+     *             }
      *         }
      *     ]
      *     results (Required): [
@@ -134,7 +156,18 @@ public final class RpcClient {
      *     status: String(InProgress/Succeeded/Failed/Canceled) (Required)
      *     errors (Optional): [
      *          (Optional){
-     *             error: ResponseError (Required)
+     *             error (Required): {
+     *                 code: String (Required)
+     *                 message: String (Required)
+     *                 target: String (Optional)
+     *                 details (Required): [
+     *                     (recursive schema, see above)
+     *                 ]
+     *                 innererror (Optional): {
+     *                     code: String (Required)
+     *                     innererror (Optional): (recursive schema, see innererror above)
+     *                 }
+     *             }
      *         }
      *     ]
      *     results (Required): [