+++ typespec-extension/changelog.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 0.7.0 (2023-05-11)
+
+Compatible with compiler 0.44.
+
+- Supported `@encode(DurationKnownEncoding.seconds, ..)` from compiler.
+
 ## 0.6.2 (2023-04-25)
 
 Compatible with compiler 0.43.